uclibc: fix build, enable CI, add unstable support of 64-bit time_t
(backport <https://github.com/rust-lang/libc/pull/5046>)
(cherry picked from commit 2549bd4a98192194ff655d0184cb7adb6152427f)

+++ src/unix/linux_like/linux/mod.rs
@@ -1268,8 +1268,9 @@ cfg_if! {
     }
 }
 
-pub const POSIX_SPAWN_USEVFORK: c_int = 64;
-pub const POSIX_SPAWN_SETSID: c_int = 128;
+pub const POSIX_SPAWN_USEVFORK: c_short = 64;
+#[cfg(not(target_env = "uclibc"))]
+pub const POSIX_SPAWN_SETSID: c_short = 128;
 
 pub const F_SEAL_FUTURE_WRITE: c_int = 0x0010;
 pub const F_SEAL_EXEC: c_int = 0x0020;
